Refactors profile page code for ease of readability.
