+++ src/Components/Profile.js
@@ -1,134 +1,42 @@
 import "../Styles/Profile.css";
-
 import React, { useContext, useEffect, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { useNavigate } from "react-router-dom";
 import { auth, db, logout } from "./Firebase";
-
 import { LocalUserContext } from "./LocalUserContext";
 import {
-  Avatar,
-  Box,
   Button,
   Container,
-  Divider,
   IconButton,
   List,
   ListItem,
-  ListItemAvatar,
-  ListItemButton,
-  ListItemText,
-  Tooltip,
   Typography,
   useTheme,
 } from "@mui/material";
 import { PopulateFromFirestore, SaveToFirestore } from "./Firestore";
-import { ArrowRight, Camera, Check, Pencil, X, XCircle } from "phosphor-react";
+import { ArrowRight, X } from "phosphor-react";
 import NoResultsImage from "./NoResultsImage";
-import ChooseAvatar from "./ChooseAvatar";
 import UserBio from "./UserBio";
-import Top8 from "./Top8List";
 import Top8List from "./Top8List";
+import ProfileList from "./ProfileList";
+import ProfileUserBanner from "./ProfileUserBanner";
 
 export default function Profile() {
   const [localUser, setLocalUser] = useContext(LocalUserContext);
-
   const [user, loading, error] = useAuthState(auth);
-  const [name, setName] = useState("");
   const navigate = useNavigate();
-  const [editAvatar, setEditAvatar] = useState(false);
 
   const theme = useTheme();
 
   useEffect(() => {
     if (loading) return;
     if (user) PopulateFromFirestore(user, localUser, setLocalUser);
-    // fetchUserName();
-    console.log(localUser);
   }, [user, loading]);
-
-  function handleAvatarToggle() {
-    setEditAvatar((editAvatar) => !editAvatar);
-  }
 
   return (
     <Container maxWidth="sm">
-      {/*****************User Banner******************/}
-      <div style={{ display: "flex", alignItems: "center" }}>
-        <Tooltip title="Change your avatar">
-          <ListItemButton
-            onClick={handleAvatarToggle}
-            sx={{ maxWidth: "110px" }}
-          >
-            <Avatar
-              sx={{ width: "80px", height: "80px" }}
-              alt={user?.displayName}
-              src={localUser?.avatar}
-            ></Avatar>{" "}
-            <Box
-              sx={{
-                position: "absolute",
-                left: "70px",
-                top: "66px",
-                width: "30px",
-                height: "30px",
-                borderRadius: "20px",
-                backgroundColor: "custom.subtleCardBg",
-                boxSizing: "border-box",
-                padding: 0,
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center",
-              }}
-            >
-              <Camera sx={{ color: "text" }} size={24} />
-            </Box>
-          </ListItemButton>
-        </Tooltip>
-
-        <ListItemText
-          sx={{ ml: 3 }}
-          primary={user?.displayName ? user.displayName : "Guest"}
-          primaryTypographyProps={{
-            fontFamily: "interBlack",
-            fontSize: "1.66rem",
-          }}
-          secondary={user?.email}
-          secondaryTypographyProps={{
-            fontFamily: "interMedium",
-            fontSize: "1rem",
-            color: "text",
-          }}
-        />
-      </div>
-
-      {editAvatar ? (
-        <Box
-          sx={{
-            backgroundColor: "custom.subtleCardBg",
-            pt: 2,
-            borderRadius: "24px",
-          }}
-        >
-          <div
-            className="row"
-            style={{ position: "relative", justifyContent: "end" }}
-          >
-            <IconButton
-              // variant="text"
-              color="inherit"
-              size="small"
-              sx={{ mr: 1, position: "absolute", top: "-5px" }}
-              onClick={handleAvatarToggle}
-            >
-              <X size={40} color="grey" />
-            </IconButton>
-          </div>
-          <ChooseAvatar />{" "}
-        </Box>
-      ) : (
-        ""
-      )}
+      {/**********************User Banner***********************/}
+      <ProfileUserBanner />
 
       {/**********************Upsell Button***********************/}
       {localUser?.name === "guest" ? (
@@ -176,45 +84,7 @@ export default function Profile() {
         {localUser && localUser["likes"]?.length > 0 ? (
           <List>
             {localUser["likes"].map((item, index) => (
-              <ListItem
-                key={index}
-                disablePadding={true}
-                disableGutters={true}
-                secondaryAction={
-                  <IconButton
-                    edge="end"
-                    aria-label="delete"
-                    onClick={() => {
-                      localUser["likes"].splice(index, 1);
-                      setLocalUser({ ...localUser });
-                      SaveToFirestore(user, localUser);
-                    }}
-                  >
-                    <X size={20} />{" "}
-                  </IconButton>
-                }
-              >
-                <ListItemButton
-                  sx={{ padding: 0 }}
-                  onClick={() => {
-                    navigate(`/anime/${item.id}`, { state: item });
-                  }}
-                >
-                  <ListItemAvatar>
-                    <Box
-                      component="img"
-                      alt={item.display_name}
-                      src={item.image_large ?? item.image_small}
-                      sx={{ height: "56px" }}
-                    ></Box>
-                  </ListItemAvatar>
-
-                  <ListItemText
-                    primary={item.display_name}
-                    primaryTypographyProps={{ fontFamily: "interMedium" }}
-                  />
-                </ListItemButton>
-              </ListItem>
+              <ProfileList item={item} index={index} />
             ))}
           </List>
         ) : (
@@ -240,44 +110,7 @@ export default function Profile() {
         {localUser && localUser["dislikes"]?.length > 0 ? (
           <List>
             {localUser["dislikes"].map((item, index) => (
-              <ListItem
-                key={index}
-                disablePadding={true}
-                disableGutters={true}
-                secondaryAction={
-                  <IconButton
-                    edge="end"
-                    aria-label="delete"
-                    onClick={() => {
-                      localUser["dislikes"].splice(index, 1);
-                      setLocalUser({ ...localUser });
-                      SaveToFirestore(user, localUser);
-                    }}
-                  >
-                    <X size={20} />{" "}
-                  </IconButton>
-                }
-              >
-                <ListItemButton
-                  sx={{ padding: 0 }}
-                  onClick={() => {
-                    navigate(`/anime/${item.id}`, { state: item });
-                  }}
-                >
-                  <ListItemAvatar>
-                    <Box
-                      component="img"
-                      alt={item.display_name}
-                      src={item.image_large ?? item.image_small}
-                      sx={{ height: "56px" }}
-                    ></Box>{" "}
-                  </ListItemAvatar>
-                  <ListItemText
-                    primary={item.display_name}
-                    primaryTypographyProps={{ fontFamily: "interMedium" }}
-                  />
-                </ListItemButton>
-              </ListItem>
+              <ProfileList item={item} index={index} />
             ))}
           </List>
         ) : (
@@ -331,49 +164,7 @@ export default function Profile() {
 
               <List>
                 {item["anime"]?.map((animeItem, animeIndex) => (
-                  <ListItem
-                    key={animeIndex}
-                    disablePadding={true}
-                    disableGutters={true}
-                    secondaryAction={
-                      <IconButton
-                        edge="end"
-                        aria-label="delete"
-                        onClick={() => {
-                          localUser["lists"][index]["anime"].splice(
-                            animeIndex,
-                            1
-                          );
-                          setLocalUser({ ...localUser });
-                          SaveToFirestore(user, localUser);
-                        }}
-                      >
-                        <X size={20} />
-                      </IconButton>
-                    }
-                  >
-                    <ListItemButton
-                      sx={{ padding: 0 }}
-                      onClick={() => {
-                        navigate(`/anime/${animeItem.id}`, {
-                          state: animeItem,
-                        });
-                      }}
-                    >
-                      <ListItemAvatar>
-                        <Box
-                          component="img"
-                          alt={animeItem.display_name}
-                          src={animeItem.image_large ?? animeItem.image_small}
-                          sx={{ height: "56px" }}
-                        ></Box>{" "}
-                      </ListItemAvatar>
-                      <ListItemText
-                        primary={animeItem.display_name}
-                        primaryTypographyProps={{ fontFamily: "interMedium" }}
-                      />
-                    </ListItemButton>
-                  </ListItem>
+                  <ProfileList item={animeItem} index={animeIndex} />
                 ))}
               </List>
             </div>
